feat: Storage Item as a dependency batched fetch (#108)
* feat: Refactor and fetch storage items batched

* misc: Add tests

* feat: Introduce resolving storage_item from context for dep_storage_item

The historic method previously would reconcile storage item on an item basis via user_client and label_row.backing_item_uuid.
For the task runner, given that the resolution algorithm has access to the context, we can actually change the previous method to basically look up the answer in the context.

Tests ammended to check that this works. We could additionally check in the tests that we don't make extraneous calls but that might be overkill

* style: Remove usage of the label row as a dependency in the various dependencies. Mid sized refactor